Move `ty::{TraitPredicate, PolyTraitPredicate}` to their own little module (cute)
